Add authorization policy for languages.

+++ app/Http/Controllers/LanguagesController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Description;
 use Illuminate\Http\Request;
-
+use Gate;
 use App\Language;
 use App\Http\Requests;
 use App\Http\Controllers\Controller;
@@ -32,6 +32,10 @@ class LanguagesController extends Controller
      */
     public function index()
     {
+//        if (Gate::denies('index')) {
+//
+//            return redirect()->back();
+//        }
         $languages = Language::orderBy('name')->get();
         return view('languages.index', compact('languages'));
     }
@@ -43,6 +47,10 @@ class LanguagesController extends Controller
      */
     public function create()
     {
+        if (Gate::denies('create')) {
+
+            return redirect()->back();
+        }
         //
         return view('languages.create');
     }
@@ -55,6 +63,10 @@ class LanguagesController extends Controller
      */
     public function store(Request $request)
     {
+        if (Gate::denies('store')) {
+
+            return redirect()->back();
+        }
         $data = $request->all();
         $newLang = Language::create($data);
         Description::create(['description' => "I'm a new language!", 'language_id' => $newLang->id]);
@@ -70,6 +82,10 @@ class LanguagesController extends Controller
     public function show($id)
     {
         $language = Language::where('id', $id)->firstOrFail();
+//        if (Gate::denies('show', $language)) {
+//
+//            return redirect()->back();
+//        }
         return view('languages.show', compact('language'));
     }
 
@@ -82,6 +98,10 @@ class LanguagesController extends Controller
     public function edit($id)
     {
         $language = Language::where('id', $id)->firstOrFail();
+        if (Gate::denies('edit', $language)) {
+
+            return redirect()->back();
+        }
         return view('languages.edit', compact('language'));
     }
 
@@ -96,6 +116,10 @@ class LanguagesController extends Controller
     {
         $data = $request->all();
         $language = Language::where('id', $id)->firstOrFail();
+        if (Gate::denies('update', $language)) {
+
+            return redirect()->back();
+        }
         $language->update($data);
 
         return redirect('languages');
@@ -110,6 +134,10 @@ class LanguagesController extends Controller
     public function destroy($id)
     {
         $language = Language::where('id', $id)->firstOrFail();
+        if (Gate::denies('destroy', $language)) {
+
+            return redirect()->back();
+        }
         $language->delete();
 
         return redirect('languages');
